requirements.txt
Flask==3.0.3
gunicorn==22.0.0
python-dotenv==1.0.1
google-cloud-aiplatform>=1.70.0
pymongo==4.6.1
Werkzeug==3.0.1